chore: update service dns
